move enabled features to common
