+++ apps/studio/hooks/misc/useIsFeatureEnabled.ts
@@ -1,54 +1,21 @@
-import type { Profile } from 'data/profile/types'
+import { isFeatureEnabled, type Feature } from 'common'
 import { useProfile } from 'lib/profile'
-import enabledFeaturesStaticObj from '../../enabled-features.json'
-
-export type Feature = Profile['disabled_features'][number] | keyof typeof enabledFeaturesStaticObj
-
-const disabledFeaturesStaticArray = Object.entries(enabledFeaturesStaticObj)
-  .filter(([_, value]) => !value)
-  .map(([key]) => key as Feature)
-
-function checkFeature(feature: Feature, features: Set<Feature>) {
-  return !features.has(feature)
-}
-
-type SnakeToCamelCase<S extends string> = S extends `${infer First}_${infer Rest}`
-  ? `${First}${SnakeToCamelCase<Capitalize<Rest>>}`
-  : S
-
-type FeatureToCamelCase<S extends Feature> = S extends `${infer P}:${infer R}`
-  ? `${SnakeToCamelCase<P>}${Capitalize<SnakeToCamelCase<R>>}`
-  : SnakeToCamelCase<S>
-
-function featureToCamelCase(feature: Feature) {
-  return feature
-    .replace(/:/g, '_')
-    .split('_')
-    .map((word, index) => (index === 0 ? word : word[0].toUpperCase() + word.slice(1)))
-    .join('') as FeatureToCamelCase<typeof feature>
-}
 
 function useIsFeatureEnabled<T extends Feature[]>(
   features: T
-): { [key in FeatureToCamelCase<T[number]>]: boolean }
-function useIsFeatureEnabled(features: Feature): boolean
-function useIsFeatureEnabled<T extends Feature | Feature[]>(features: T) {
+): ReturnType<typeof isFeatureEnabled<T>>
+function useIsFeatureEnabled(features: Feature): ReturnType<typeof isFeatureEnabled>
+function useIsFeatureEnabled<T extends Feature | Feature[]>(features: T): any {
   const { profile } = useProfile()
-  const disabledFeatures = new Set([
-    ...(profile?.disabled_features ?? []),
-    ...disabledFeaturesStaticArray,
-  ])
+  const disabledFeatures = profile?.disabled_features
 
+  // This code branch is to make the type checker happy, it's intentionally
+  // the same as the isFeatureEnabled function call below.
   if (Array.isArray(features)) {
-    return Object.fromEntries(
-      features.map((feature) => [
-        featureToCamelCase(feature),
-        checkFeature(feature, disabledFeatures),
-      ])
-    )
+    return isFeatureEnabled(features, disabledFeatures)
   }
 
-  return checkFeature(features, disabledFeatures)
+  return isFeatureEnabled(features, disabledFeatures)
 }
 
 export { useIsFeatureEnabled }
