+++ packages/common/index.tsx
@@ -2,10 +2,11 @@ export * from './auth'
 export * from './consent-state'
 export * from './constants'
 export * from './database-types'
+export * from './enabled-features'
+export * from './feature-flags'
 export * from './gotrue'
 export * from './helpers'
 export * from './hooks'
 export * from './MetaFavicons/pages-router'
 export * from './Providers'
 export * from './telemetry'
-export * from './feature-flags'
